refactore(#20): 알림 서비스 레이어 의존성 역전
